refactor(api): extract storage ports to types-storage.ts (#978) (#1001)
repositories/types.ts sat pinned at the 800-line hard cap, so the next
interface addition would force a size-driven extraction onto an unrelated
contributor. Carve the two object-storage ports (PhotoStorage,
DocumentStorage) into repositories/types-storage.ts and re-export from the
barrel, matching the existing types-payment.ts / types-notification.ts seam.

No behavior change — every consumer still imports from ./types. types.ts
drops to 789 lines (11 lines of headroom).
