+++ packages/api/src/repositories/types.ts
@@ -72,6 +72,11 @@ export type {
   NotificationLogRepository,
   NotificationLogUpsert,
 } from './types-notification'
+
+// Object-storage ports (PhotoStorage #461/#952, DocumentStorage #459) live in
+// their own module to keep this barrel under the file-size cap (#978);
+// re-exported for callers.
+export type { DocumentStorage, PhotoStorage } from './types-storage'
 export { complianceAlertKey } from './types-compliance'
 export type { ComplianceAlertLogRepository, RecordComplianceAlert } from './types-compliance'
 
@@ -738,12 +743,6 @@ export interface FeeScheduleRepository {
   archive(id: string): Promise<FeeSchedule | undefined>
 }
 
-export interface PhotoStorage {
-  put(vehicleId: string, file: File): Promise<{ key: string; url: string }>
-  /** Accepts a key or full URL. `ownerVehicleId` scopes the delete to that vehicle's key prefix, so a cross-referenced URL can't reach another vehicle's object (cross-tenant IDOR guard, #952). */
-  delete(keyOrUrl: string, ownerVehicleId: string): Promise<void>
-}
-
 export interface RenterDocumentFilters {
   limit?: number
   offset?: number
@@ -788,13 +787,3 @@ export interface RenterDocumentRepository {
 }
 
 export type CreateRenterDocumentData = Pick<RenterDocument, 'renterId' | 'type' | 'storageKey'>
-
-/** Private object storage for renter document scans (#459) — private, not public-URL addressable (cf. `PhotoStorage`). */
-export interface DocumentStorage {
-  put(renterId: string, file: File): Promise<{ key: string }>
-  /** @deprecated Dormant pending #304; superseded by getFile(). Signed URL to view a scan. */
-  getSignedUrl(key: string): Promise<string>
-  /** Stream a stored scan's bytes + content-type for an authenticated viewer (#932). */
-  getFile(key: string): Promise<{ body: ReadableStream; contentType: string } | null>
-  delete(key: string): Promise<void>
-}
